move 'is_new_or_updated()' to 'lib/common_funcs.py'
